Add files via upload
Machine learning algorithms can predict the cost of the premium for health insurance. The cost of the premium for health insurance varies from person to person. For example, a smoker is charged a higher premium than a non-smoker. Companies calculate the premium for each person based on a few factors like age, BMI, smoker, etc. It is possible to predict the premium cost using machine learning algorithms. We can train a machine learning algorithm using past data to predict the premium cost.
The size of past data is 5352 records. It contained 7 columns, out of which six of them was the independent variable and one was the dependent variable. The data was labeled so we can do supervised learning algorithms. As the dependent variable is a continuous numerical value this is a regression problem. The data contains no missing values but it does contain outliers.
However, some machine learning algorithms like a decision tree and random forest are not sensitive to outliers, therefore we have not removed the outliers. The variables have a good correlation except for regions. The variables like BMI, children, and smokers are normally distributed but the others are not. Feature scaling using a min-max scaler was done to give all features equal importance. Regions were removed as they had no correlation. Different ml algorithms were trained using the dataset and tested.
The decision tree gave the best r2 score of 97.56% and an adjusted r2 score of 97.56% as well. We trained the decision tree algorithm again with the best parameters from randomized search. The r2 score increased to 98.401% and the adjusted r2 score was 98.399%. The r2 score is greater than the adjusted r2 score. A visualization of the tree was also done to have a better interpretation of the algorithm.